+++ Service/Product/IndexOptionsBuilder.php
@@ -2,18 +2,19 @@
 
 namespace Algolia\AlgoliaSearch\Service\Product;
 
+use Algolia\AlgoliaSearch\Api\Builder\EntityIndexOptionsBuilderInterface;
 use Algolia\AlgoliaSearch\Api\Data\IndexOptionsInterface;
 use Algolia\AlgoliaSearch\Helper\Entity\ProductHelper;
-use Algolia\AlgoliaSearch\Service\AbstractEntityIndexOptionsBuilder;
+use Algolia\AlgoliaSearch\Service\IndexOptionsBuilder as BaseIndexOptionsBuilder;
 use Magento\Framework\Exception\NoSuchEntityException;
 
-class IndexOptionsBuilder extends AbstractEntityIndexOptionsBuilder
+class IndexOptionsBuilder extends BaseIndexOptionsBuilder implements EntityIndexOptionsBuilderInterface
 {
     /**
      * @throws NoSuchEntityException
      */
     public function buildEntityIndexOptions(int $storeId, ?bool $isTmp = false): IndexOptionsInterface
     {
-        return $this->safeBuildWithComputedIndex(ProductHelper::INDEX_NAME_SUFFIX, $storeId, $isTmp);
+        return $this->buildWithComputedIndex(ProductHelper::INDEX_NAME_SUFFIX, $storeId, $isTmp);
     }
 }